+++ src/components/buttons/z-chip/index.spec.ts
@@ -53,7 +53,7 @@ describe("Suite test ZChip", () => {
 
     expect(page.root).toEqualHtml(`
       <z-chip icon="" interactive-icon="multiply-circled" type="default">
-          <button class="default" tabindex="0">
+          <button class="default" tabindex="0" type="button">
             text
             <z-icon height="14" name="multiply-circled" tabindex="0" width="14"></z-icon>
           </button>
@@ -69,7 +69,7 @@ describe("Suite test ZChip", () => {
 
     expect(page.root).toEqualHtml(`
       <z-chip icon="pdf" interactive-icon="multiply-circled" type="default">
-          <button class="default" tabindex="0">
+          <button class="default" tabindex="0" type="button">
             <z-icon class="iconSx" height="14" name="pdf" width="14"></z-icon>
             text
             <z-icon height="14" name="multiply-circled" tabindex="0" width="14"></z-icon>
@@ -86,7 +86,7 @@ describe("Suite test ZChip", () => {
 
     expect(page.root).toEqualHtml(`
       <z-chip disabled="" icon="pdf" interactive-icon="multiply-circled" type="default">
-          <button class="default" tabindex="0" disabled>
+          <button class="default" tabindex="0" disabled type="button">
             <z-icon class="iconSx" height="14" name="pdf" width="14"></z-icon>
             text
             <z-icon height="14" name="multiply-circled" tabindex="-1" width="14"></z-icon>
@@ -103,7 +103,7 @@ describe("Suite test ZChip", () => {
 
     expect(page.root).toEqualHtml(`
       <z-chip icon="pdf" interactive-icon="multiply-circled" type="small">
-          <button class="small" tabindex="0">
+          <button class="small" tabindex="0" type="button">
             <z-icon class="iconSx" height="14" name="pdf" width="14"></z-icon>
             text
             <z-icon height="14" name="multiply-circled" tabindex="0" width="14"></z-icon>